feat(main.tf): add Terraform configuration file to provision an AWS EC2 instance
The main.tf file is added to the project, which contains the Terraform configuration for provisioning an AWS EC2 instance. The configuration specifies the required provider, AWS, with the profile set to "default" and the region set to "us-east-1". It also defines an AWS instance resource with the AMI set to "ami-0230bd60aa48260c6" and the instance type set to "t2.micro". Additionally, a tag with the name "ExampleAppServerInstance" is added to the instance.