feat(ec2_instance): add Terraform configuration for creating an AWS EC2 instance
This commit adds a new file `main.tf` in the `ec2_instance` directory. The `main.tf` file contains the Terraform configuration for creating an AWS EC2 instance.

The configuration specifies the required provider `aws` with a version constraint of `~> 5.0`. It also sets the required Terraform version to be `>= 1.2.0`.

The `aws_instance` resource is defined with the AMI ID `ami-0230bd60aa48260c6` and the instance type `t2.micro`. It also includes a tag with the name "ExampleAppServerInstance".

This configuration can be used to provision an EC2 instance on AWS using Terraform.